requirements.txt
agora-token-builder==1.0.0
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
awsebcli==3.20.3
beautifulsoup4==4.11.2
botocore==1.23.54
bs4==0.0.1
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
DateTime==5.0
Django==3.2.17
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-cors-headers==3.14.0
django-honeypot==1.0.3
django-session-timeout==0.1.0
exceptiongroup==1.1.0
future==0.16.0
h11==0.14.0
idna==3.4
jmespath==0.10.0
mysql-connector-python==8.0.32
numpy==1.21.6
outcome==1.2.0
pandas==1.1.5
pathspec==0.9.0
Pillow==9.4.0
protobuf==3.20.3
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7
PyYAML==5.4.1
requests==2.26.0
selenium==4.7.0
semantic-version==2.8.5
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
termcolor==1.1.0
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.5
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.1.9
wsproto==1.2.0
zope.interface==5.5.2
